Arreglos
Arreglo del componente producto, validación de login, arreglo de siderbar y quitar los botones de header.

+++ src/app/pages/crud-proveedor/crud-proveedor.component.ts
@@ -37,10 +37,13 @@ export class CrudProveedorComponent implements OnInit {
   }
 
   eliminar(proveedor: Proveedor): void {
-    this.proveedorService.deleteProveedor(proveedor)
-      .subscribe(data => {
-        this.proveedor = this.proveedor.filter(p => p !== proveedor);
-        alert("Se elimino...!!")
-      })
+    if (confirm('¿Seguro deseas eliminar este proveedor?')) {
+      this.proveedorService.deleteProveedor(proveedor)
+        .subscribe(data => {
+          this.proveedor = this.proveedor.filter(p => p !== proveedor);
+          alert("Se elimino...!!")
+        });
+    }
+
   }
 }
